Resolved CID 351553 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument World::getInstance()->getSessionCount() to format specifier %u was expected to have type unsigned int but has type unsigned long

+++ src/world/Server/Packets/Handlers/CharacterHandler.cpp
@@ -620,7 +620,7 @@ void WorldSession::sendServerStats()
 
         _player->BroadcastMessage("Build hash: %s%s", MSG_COLOR_CYAN, BUILD_HASH_STR);
         _player->BroadcastMessage("Online Players: %s%u |rPeak: %s%u|r Accepted Connections: %s%u",
-            MSG_COLOR_SEXGREEN, sWorld.getSessionCount(), MSG_COLOR_SEXBLUE, sWorld.getPeakSessionCount(),
+            MSG_COLOR_SEXGREEN, static_cast<uint32_t>(sWorld.getSessionCount()), MSG_COLOR_SEXBLUE, sWorld.getPeakSessionCount(),
             MSG_COLOR_SEXBLUE, sWorld.getAcceptedConnections());
 
         _player->BroadcastMessage("Server Uptime: |r%s", sWorld.getWorldUptimeString().c_str());
